requirements.txt
fastapi==0.109.1
uvicorn==0.27.0.post1
sqlalchemy>=2.0.0,<2.1.0
pydantic>=2.0.0
python-multipart==0.0.6
langchain==0.1.1
chromadb==0.4.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sentence-transformers==2.5.1
transformers==4.37.2
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
accelerate==0.27.2
bitsandbytes==0.42.0
